[REFACTOR] BaseBottomSheetFragment 추가 및 적용 (#19)
