+++ common-ui/src/main/java/com/lgtm/android/common_ui/base/BaseBottomSheetFragment.kt
@@ -1,0 +1,56 @@
+package com.lgtm.android.common_ui.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.android.material.bottomsheet.BottomSheetDialog
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+
+abstract class BaseBottomSheetFragment<T : ViewDataBinding>(
+    @LayoutRes private val layoutRes: Int
+) : BottomSheetDialogFragment() {
+    private var _binding: T? = null
+    protected val binding: T
+        get() = requireNotNull(_binding)
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
+        binding.lifecycleOwner = this
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        setBottomSheetBehavior()
+    }
+
+    private fun setBottomSheetBehavior() {
+        (dialog as BottomSheetDialog).behavior.apply {
+            state = BottomSheetBehavior.STATE_EXPANDED
+            skipCollapsed = true
+            isDraggable = false
+        }
+    }
+
+    open fun setBottomSheetHeight(ratio: Double) {
+        binding.root.layoutParams.height =
+            (resources.displayMetrics.heightPixels * ratio).toInt()
+    }
+
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+
+
+}
